+++ src/code-gen-process.js
@@ -81,6 +81,7 @@ class CodeGenProcess {
       this.templates,
       this.typeName,
     );
+    this.config.componentTypeNameResolver.logger = this.logger;
   }
 
   async start() {
@@ -176,9 +177,11 @@ class CodeGenProcess {
 
     if (this.fileSystem.pathIsExist(this.config.output)) {
       if (this.config.cleanOutput) {
+        this.logger.debug(`cleaning dir ${this.config.output}`);
         this.fileSystem.cleanDir(this.config.output);
       }
     } else {
+      this.logger.debug(`path ${this.config.output} is not exist. creating dir by this path`);
       this.fileSystem.createDir(this.config.output);
     }
 
@@ -403,15 +406,11 @@ class CodeGenProcess {
     if (configuration.translateToJavaScript) {
       const { sourceContent, declarationContent } = translateToJS(`${fixedFileName}${ts.Extension.Ts}`, content);
 
-      if (this.config.debug) {
-        console.info("generating output for", `${fixedFileName}${ts.Extension.Js}`);
-        console.info(sourceContent);
-      }
+      this.logger.debug("generating output for", `${fixedFileName}${ts.Extension.Js}`);
+      this.logger.debug(sourceContent);
 
-      if (this.config.debug) {
-        console.info("generating output for", `${fixedFileName}${ts.Extension.Dts}`);
-        console.info(declarationContent);
-      }
+      this.logger.debug("generating output for", `${fixedFileName}${ts.Extension.Js}`);
+      this.logger.debug(declarationContent);
 
       return {
         name: `${fixedFileName}${ts.Extension.Js}`,
@@ -423,10 +422,8 @@ class CodeGenProcess {
       };
     }
 
-    if (this.config.debug) {
-      console.info("generating output for", `${fixedFileName}${ts.Extension.Ts}`);
-      console.info(content);
-    }
+    this.logger.debug("generating output for", `${fixedFileName}${ts.Extension.Js}`);
+    this.logger.debug(content);
 
     return {
       name: `${fixedFileName}${ts.Extension.Ts}`,