+++ src/schema-parser/schema-parser.js
@@ -105,6 +105,19 @@ class SchemaParser {
 
       const refType = this.schemaUtils.getSchemaRefType(schema);
       const $ref = (refType && refType.$ref) || null;
+
+      if (Array.isArray(schema.enum) && Array.isArray(schema.enum[0])) {
+        return this.parseSchema(
+          {
+            oneOf: schema.enum.map((enumNames) => ({
+              type: "array",
+              items: enumNames.map((enumName) => ({ type: "string", enum: [enumName] })),
+            })),
+          },
+          typeName,
+        );
+      }
+
       const keyType = await this.getSchemaType(schema);
       const enumNames = this.schemaUtils.getEnumNames(schema);
       let content = null;
@@ -221,7 +234,7 @@ class SchemaParser {
     },
     [SCHEMA_TYPES.PRIMITIVE]: async (schema, typeName) => {
       let contentType = null;
-      const { additionalProperties, type, description, $$requiredKeys } = schema || {};
+      const { additionalProperties, type, description, items } = schema || {};
 
       if (type === this.config.Ts.Keyword.Object && additionalProperties) {
         const fieldType = _.isObject(additionalProperties)
@@ -235,6 +248,10 @@ class SchemaParser {
           ...(_.isObject(schema) ? schema : {}),
           oneOf: type.map((type) => ({ type })),
         });
+      }
+
+      if (_.isArray(items) && type === SCHEMA_TYPES.ARRAY) {
+        contentType = this.config.Ts.Tuple(items.map((item) => this.getInlineParseContent(item)));
       }
 
       return {
@@ -379,7 +396,7 @@ class SchemaParser {
         typeName = await this.getSchemaType(schema);
       }
 
-      if (schema.items && !schema.type) {
+      if (schema.items && !Array.isArray(schema.items) && !schema.type) {
         schema.type = SCHEMA_TYPES.ARRAY;
       }
       schemaType = this.getInternalSchemaType(schema);